switch to deadline mode to avoid generating false positive error reports in an unreliable network.
now we send an icmp echo request and wait for a response with a
timeout. if the packet is discarded somewhere in the net we treat
it like the address unreachable problem though a reasonable packet
loss is acceptable actually.
let's switch to the deadline mode where we keep sending requests
for a reasonable timeout and wait for a reply for any of them.